requirements.txt
altgraph==0.17.4
asttokens==2.4.1
blessings==1.7
blinker==1.7.0
bottle==0.12.25
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
colour-runner==0.1.1
comm==0.2.2
coverage==7.4.2
debugpy==1.8.1
decorator==5.1.1
deepdiff==6.7.1
distlib==0.3.8
executing==2.0.1
filelock==3.13.1
Flask==3.0.2
greenlet==3.0.3
humanfriendly==10.0
idna==3.6
ipykernel==6.29.4
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
ordered-set==4.1.0
packaging==23.2
pandas==2.2.0
parso==0.8.4
pefile==2023.2.7
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
proxy_tools==0.1.0
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
pyproject-api==1.6.1
pyreadline3==3.4.1
python-dateutil==2.8.2
pythonnet==3.0.3
pytz==2024.1
pywebview==4.4.1
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==26.0.2
requests==2.31.0
rootpath==0.1.1
setuptools==69.1.0
six==1.16.0
SQLAlchemy==2.0.27
stack-data==0.6.3
ta==0.11.0
termcolor==2.4.0
tornado==6.4
tox==4.13.0
traitlets==5.14.3
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.0
wcwidth==0.2.13
Werkzeug==3.0.1
